Extract V1 specific logic out of ApiController
